effit-backend/src/main/kotlin/be/swsb/effit/challenge/ChallengeController.kt
package be.swsb.effit.challenge

import org.springframework.http.MediaType
import org.springframework.http.ResponseEntity
import org.springframework.web.bind.annotation.GetMapping
import org.springframework.web.bind.annotation.RequestMapping
import org.springframework.web.bind.annotation.RestController

@RestController
@RequestMapping("/api/challenge",
        produces = [MediaType.APPLICATION_JSON_UTF8_VALUE])
class ChallengeController(private val challengeRepository: ChallengeRepository) {

    @GetMapping
    fun allChallenges() : ResponseEntity<List<Challenge>> {
        return ResponseEntity.ok(challengeRepository.findAll())
    }
}